loss tests + double -> float in loss functions

+++ src/LossFunctions/BinaryCrossEntropy.cpp
@@ -2,7 +2,7 @@
 
 namespace neural_net {
 
-double BinaryCrossEntropy::Loss(const Matrix& present, const Matrix& expected) const {
+float BinaryCrossEntropy::Loss(const Matrix& present, const Matrix& expected) const {
     assert(present.rows() == expected.rows());
     assert(present.cols() == 1 && expected.cols() == 1 &&
            "it should be output layer size equal 1 for binarycrossentropy loss function");
